avoid duplicate defining of register component
