+++ examples/basic-form/src/main.js
@@ -12,48 +12,16 @@ import TextBox from './components/TextBox';
 import Button from './components/Button';
 
 
-const styles = StyleSheet.create({
-  register: {
-    backgroundColor: colors.LightGrey,
-    padding: spacing.Large
-  },
-  heading: {
-    color: colors.Purple,
-    fontSize: typeRamp.Medium,
-    fontFamily: fontFamily,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    marginBottom: spacing.Medium,
-    width: 300
-  }
-});
+
 
 const Page = ({ sessions }: { sessions: Array<Session> }) => (
   <View>
     <Text style={typography.Heading}>Basic Form w/ DOM elements and React Primitives</Text>
-    <Register>
-      {sessions.map(session => (
-        <Space key={session.password} h={spacing.Large} v={spacing.Large}>
-          <View style={styles.register}>
-            <Text style={styles.heading}>Register an Account</Text>
-            <TextBox
-              label={"Email"}
-              value={session.email}
-              type={"email"}
-            />
-            <TextBox
-              label={"Password"}
-              value={session.password}
-              type={"password"}
-            />
-            <Button
-              label={"Register"}
-              backgroundColor={colors.Purple}
-            />
-          </View>
-        </Space>
-      ))}
-    </Register>
+    {sessions.map(session => (
+      <Register 
+        session={session}
+      />
+    ))}
   </View>
 );
