+++ examples/basic-form/src/web.js
@@ -2,9 +2,7 @@ import { Text } from 'react-primitives';
 import React from 'react';
 import { colors, spacing, typeRamp, typography, fontFamily } from './designSystem';
 import Register from './components/Register';
-import Space from './components/Space';
-import TextBox from './components/TextBox';
-import Button from './components/Button';
+
 
 
 const styles = {
@@ -12,20 +10,6 @@ const styles = {
     width: 364,
     marginLeft: 'auto',
     marginRight: 'auto'
-  },
-  register: {
-    backgroundColor: colors.LightGrey,
-    padding: spacing.Large
-  },
-  heading: {
-    color: colors.Purple,
-    display: 'block',
-    fontSize: typeRamp.Medium,
-    fontFamily: fontFamily,
-    fontWeight: 'bold',
-    textAlign: 'center',
-    marginBottom: spacing.Medium,
-    width: 300
   }
 };
 
@@ -38,25 +22,8 @@ export default () => (
           ...typography.Heading,
           fontFamily: fontFamily
         }}>Basic Form w/ DOM elements and React Primitives. Type a password! 👀</h1>
-      <Register>
-        <div style={styles.register}>
-          <Text style={{...styles.heading}}>Register an Account</Text>
-          <TextBox
-            label={"Email"}
-            value={""}
-            type={"email"}
-          />
-          <TextBox
-            label={"Password"}
-            value={""}
-            type={"password"}
-          />
-          <Button
-            label={"Register"}
-            backgroundColor={colors.Purple}
-          />
-        </div>
-      </Register>
+      <Register
+        session={{email: '', password: ''}}/>
     </div>
   </div>
 );
